+++ packages/ui/src/components/layouts/forms/index.ts
@@ -1,12 +1,14 @@
-import HSimpleCenterFormLayout from './HSimpleCenterFormLayout.vue';
-import HCenterFormLayout from './HCenterFormLayout.vue';
-import HColumnFormLayout from './HColumnFormLayout.vue';
-import HDetailContainer from './HDetailContainer.vue';
-import HErrorLayout from './HErrorLayout.vue';
-import HFullWidthFormLayout from './HFullWidthFormLayout.vue';
+import HSimpleCenterFormLayout from "./HSimpleCenterFormLayout.vue";
+import HCenterFormLayout from "./HCenterFormLayout.vue";
+import HColumnFormLayout from "./HColumnFormLayout.vue";
+import HDetailContainer from "./HDetailContainer.vue";
+import HErrorLayout from "./HErrorLayout.vue";
+import HInformationFormLayout from "./HInformationFormLayout.vue";
+import HFullWidthFormLayout from "./HFullWidthFormLayout.vue";
 
 export {
   HSimpleCenterFormLayout,
+  HInformationFormLayout,
   HCenterFormLayout,
   HColumnFormLayout,
   HDetailContainer,