+++ packages/ui/src/views/pages/iot/mqtt-account/Index.vue
@@ -16,7 +16,7 @@
       @request="findItems"
     >
       <template #top-left>
-        <h-button color="primary" label="新建角色" @click="toCreate" />
+        <h-button color="primary" icon="mdi-plus" label="新建角色" @click="toCreate" />
       </template>
 
       <template #body-cell-superUser="props">