+++ packages/ui/src/views/pages/security/role/Index.vue
@@ -14,7 +14,7 @@
     @request="findItems"
   >
     <template #top-left>
-      <h-button color="primary" label="新建角色" @click="toCreate" />
+      <h-button color="primary" icon="mdi-plus" label="新建角色" @click="toCreate" />
     </template>
 
     <template #body-cell-actions="props">
@@ -33,12 +33,12 @@
 </template>
 
 <script setup lang="ts">
-import type { SysRoleEntity, SysRoleConditions, SysRoleProps } from "@herodotus/api";
-import type { QTableColumnProps } from "@/composables/declarations";
+import type { SysRoleEntity, SysRoleConditions, SysRoleProps } from '@herodotus/api';
+import type { QTableColumnProps } from '@/composables/declarations';
 
-import { PAGE_NAME, API } from "@/configurations";
-import { useTable, useDateTime } from "@/composables/hooks";
-import { HDeleteButton, HEditButton, HDenseIconButton, HTable } from "@/components";
+import { PAGE_NAME, API } from '@/configurations';
+import { useTable, useDateTime } from '@/composables/hooks';
+import { HDeleteButton, HEditButton, HDenseIconButton, HTable } from '@/components';
 
 defineOptions({ name: PAGE_NAME.SYS_ROLE, components: { HDeleteButton, HEditButton, HDenseIconButton, HTable } });
 
@@ -47,27 +47,27 @@ const { tableRows, totalPages, pagination, loading, toEdit, toCreate, toAuthoriz
   useTable<SysRoleConditions, SysRoleEntity>(API.core.sysRole(), PAGE_NAME.SYS_ROLE);
 
 const selected = ref([]);
-const rowKey: SysRoleProps = "roleId";
+const rowKey: SysRoleProps = 'roleId';
 
 const columns: QTableColumnProps = [
-  { name: "roleName", field: "roleName", align: "center", label: "角色名称" },
-  { name: "roleCode", field: "roleCode", align: "center", label: "角色代码" },
-  { name: "description", field: "description", align: "center", label: "备注" },
+  { name: 'roleName', field: 'roleName', align: 'center', label: '角色名称' },
+  { name: 'roleCode', field: 'roleCode', align: 'center', label: '角色代码' },
+  { name: 'description', field: 'description', align: 'center', label: '备注' },
   {
-    name: "updateBy",
-    field: "updateBy",
-    align: "center",
-    label: "最后修改人",
+    name: 'updateBy',
+    field: 'updateBy',
+    align: 'center',
+    label: '最后修改人',
   },
   {
-    name: "updateTime",
-    field: "updateTime",
-    align: "center",
-    label: "修改时间",
+    name: 'updateTime',
+    field: 'updateTime',
+    align: 'center',
+    label: '修改时间',
     format: (value) => defaultFormat(value),
   },
-  { name: "reserved", field: "reserved", align: "center", label: "保留数据" },
-  { name: "status", field: "status", align: "center", label: "状态" },
-  { name: "actions", field: "actions", align: "center", label: "操作" },
+  { name: 'reserved', field: 'reserved', align: 'center', label: '保留数据' },
+  { name: 'status', field: 'status', align: 'center', label: '状态' },
+  { name: 'actions', field: 'actions', align: 'center', label: '操作' },
 ];
 </script>